requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
qrcode==7.4.2
gunicorn==21.2.0
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==10.0.1
cryptography==41.0.7
Werkzeug==2.3.7